refactor(coding-agent): share standalone context-file walker between providers
